Fixed DRA input tests
