+++ planscore/preread_followup.py
@@ -23,18 +23,12 @@ import osgeo.ogr
 from . import util, data, score, website, prepare_state, constants, observe
 
 FUNCTION_NAME = os.environ.get('FUNC_NAME_PREREAD_FOLLOWUP') or 'PlanScore-PrereadFollowup'
-EMPTY_GEOMETRY = osgeo.ogr.Geometry(osgeo.ogr.wkbGeometryCollection)
-POLYGONAL_TYPES = {osgeo.ogr.wkbPolygon, osgeo.ogr.wkbMultiPolygon}
 
 Assignment = collections.namedtuple('Assignment', ('block_id', 'district_id'))
 
 osgeo.ogr.UseExceptions()
 
 states_path = os.path.join(os.path.dirname(__file__), 'geodata', 'cb_2013_us_state_20m.geojson')
-
-def is_polygonal_feature(feature):
-    geometry = feature.GetGeometryRef() or EMPTY_GEOMETRY
-    return bool(geometry.GetGeometryType() in POLYGONAL_TYPES)
 
 def ordered_districts(layer):
     ''' Return field name and list of layer features ordered by guessed district numbers.
@@ -43,7 +37,7 @@ def ordered_districts(layer):
     expected_values = {i+1 for i in range(len(layer))}
     fields = list()
     
-    polygon_features = [feat for feat in layer if is_polygonal_feature(feat)]
+    polygon_features = [feat for feat in layer if util.is_polygonal_feature(feat)]
 
     for index in range(defn.GetFieldCount()):
         name = defn.GetFieldDefn(index).GetName()
@@ -199,7 +193,7 @@ def guess_geometry_model(path):
         else:
             return a.Union(b)
     
-    features = [feat for feat in ds.GetLayer(0) if is_polygonal_feature(feat)]
+    features = [feat for feat in ds.GetLayer(0) if util.is_polygonal_feature(feat)]
     geometries = [_simplify_coarsely(feature.GetGeometryRef()) for feature in features]
     footprint = functools.reduce(_union_safely, geometries)
     
@@ -293,7 +287,7 @@ def put_geojson_file(s3, bucket, upload, path):
     _, features = ordered_districts(ds.GetLayer(0))
     
     for (index, feature) in enumerate(features):
-        if not is_polygonal_feature(feature):
+        if not util.is_polygonal_feature(feature):
             continue
         geometry = feature.GetGeometryRef()
         if geometry.GetSpatialReference():
